update ( fazer alterações )

+++ src/genre/genre.service.ts
@@ -1,10 +1,12 @@
 import {
   Injectable,
   NotFoundException,
+  UnauthorizedException,
   UnprocessableEntityException,
 } from '@nestjs/common';
 import { Genre } from '@prisma/client';
 import { PrismaService } from 'src/prisma/prisma.service';
+import { User } from 'src/user/entities/user.entity';
 import { handleError } from 'src/utilis/handle-error.utilis';
 import { CreateGenreDto } from './dto/create-genre.dto';
 import { UpdateGenreDto } from './dto/update-genre.dto';
@@ -12,7 +14,7 @@ import { Gender } from './entities/genre.entity';
 
 @Injectable()
 export class GenreService {
-  constructor(private readonly prisma: PrismaService) {}
+  constructor(private readonly prisma: PrismaService) { }
 
   async findById(name: string): Promise<Genre> {
     const record = await this.prisma.genre.findUnique({ where: { name } });
@@ -32,7 +34,8 @@ export class GenreService {
     return this.prisma.genre.findMany();
   }
 
-  async update(name: string, dto: UpdateGenreDto): Promise<Genre> {
+  async update(user:User, name: string, dto: UpdateGenreDto): Promise<Genre> {
+    if (user.Admin) {
     await this.findById(name);
     const data: Partial<Genre> = { ...dto };
 
@@ -42,17 +45,29 @@ export class GenreService {
         data,
       })
       .catch(handleError);
+    }else {
+      throw new UnauthorizedException('O usuário não é um admin, contate-o o admin para acessar essa informação !')
+    }
   }
 
-  create(dto: CreateGenreDto): Promise<Genre> {
-    const genre: Gender = { ...dto };
+  create(user: User, dto: CreateGenreDto): Promise<Genre> {
+    if (user.Admin) {
+      const genre: Gender = { ...dto };
 
-    return this.prisma.genre.create({ data: genre }).catch(handleError);
+      return this.prisma.genre.create({ data: genre }).catch(handleError);
+    } else {
+      throw new UnauthorizedException('O usuário não é um admin, contate-o o admin para acessar essa informação !')
+    }
   }
 
-  async delete(name: string) {
-    await this.findById(name);
+  async delete(user:User, name: string) {
+    if (user.Admin) {
+      await this.findById(name);
 
-    await this.prisma.genre.delete({ where: { name } });
+      await this.prisma.genre.delete({ where: { name } });
+    } else {
+      throw new UnauthorizedException('O usuário não é um admin, contate-o o admin para acessar essa informação !')
+    }
   }
+
 }
